+++ swift_algorithms/Views/Catalyst/CatalystSearchViewController.swift
@@ -1,0 +1,41 @@
+import UIKit
+import Anchorage
+
+final class CatalystSearchViewController: UIViewController {
+    private let algorithmSearchResultsController = SearchResultsTableViewController(style: .grouped)
+    private let algorithmSearchResultsPresenter = AlgorithmSearchPresenter()
+    private let algorithmSearchController = UISearchController(searchResultsController: nil)
+
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        algorithmSearchResultsPresenter.deliver = { [weak algorithmSearchResultsController] properties in
+            algorithmSearchResultsController?.properties = properties
+        }
+        
+        algorithmSearchResultsPresenter.makeSearchableAlgorithmProperties()
+//        algorithmSearchResultsPresenter.dispatcher = self
+        
+        algorithmSearchResultsController.dispatcher = algorithmSearchResultsPresenter
+        
+        algorithmSearchController.searchResultsUpdater = algorithmSearchResultsController
+        algorithmSearchController.obscuresBackgroundDuringPresentation = false
+        algorithmSearchController.searchBar.placeholder = "Search Algorithms"
+        
+        navigationItem.searchController = algorithmSearchController
+        navigationItem.hidesSearchBarWhenScrolling = false
+        
+        algorithmSearchResultsController.willMove(toParent: self)
+        view.addSubview(algorithmSearchResultsController.view)
+        algorithmSearchResultsController.view.edgeAnchors == view.edgeAnchors
+        algorithmSearchResultsController.didMove(toParent: self)
+        
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        algorithmSearchController.isActive = true
+        algorithmSearchResultsController.tableView.contentInset = .init(top: 100, left: 0, bottom: 0, right: 0)
+    }
+}
