+++ swift_algorithms/Views/Catalyst/RootCatalystCoordinator.swift
@@ -40,46 +40,6 @@ final class RootCatalystCoordinator: Coordinating {
     }
 }
 
-final class CatalystSearchViewController: UIViewController {
-    
-    private let algorithmSearchResultsController = SearchResultsTableViewController(style: .grouped)
-    private let algorithmSearchResultsPresenter = AlgorithmSearchPresenter()
-    let algorithmSearchController = UISearchController(searchResultsController: nil)
-
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        algorithmSearchResultsPresenter.deliver = { [weak algorithmSearchResultsController] properties in
-            algorithmSearchResultsController?.properties = properties
-        }
-        
-        algorithmSearchResultsPresenter.makeSearchableAlgorithmProperties()
-//        algorithmSearchResultsPresenter.dispatcher = self
-        
-        algorithmSearchResultsController.dispatcher = algorithmSearchResultsPresenter
-        
-        algorithmSearchController.searchResultsUpdater = algorithmSearchResultsController
-        algorithmSearchController.obscuresBackgroundDuringPresentation = false
-        algorithmSearchController.searchBar.placeholder = "Search Algorithms"
-        
-        navigationItem.searchController = algorithmSearchController
-        navigationItem.hidesSearchBarWhenScrolling = false
-        
-        algorithmSearchResultsController.willMove(toParent: self)
-        view.addSubview(algorithmSearchResultsController.view)
-        algorithmSearchResultsController.view.edgeAnchors == view.edgeAnchors
-        algorithmSearchResultsController.didMove(toParent: self)
-        
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        algorithmSearchController.isActive = true
-        algorithmSearchResultsController.tableView.contentInset = .init(top: 100, left: 0, bottom: 0, right: 0)
-    }
-}
-
 @available(iOS 13.0, *)
 extension RootCatalystCoordinator: MainCatalystPresenterDelegate {
     func show(puzzle: Puzzle) {
